Log relay state
Relay state is logged in three places:

 * Coming from Okta to the saml proxy (the first call the saml proxy
   sees in the login flow)
 * Coming from the IDP back to the saml proxy (after the user has logged
   in)
 * Going from the saml proxy back to Okta (This log takes place in the
   same call as the second bullet point). There is potential that this
   relay state is different than what it was coming from the IDP.

There is a session Id that is logged along side the relay state to help
match up requests. At first I was concerned with the security of logging
this session ID but it is already logged all over the saml-proxy so this
at least doesn't increase any risk (might want to make sure we want that
session id logged though). All the information is split into fields in a
JSON object to help querying against it. An iso8601 datetime was added as
well.

I increased a session's max age to help make sure we keep the same
session for the same login process. Some users have to enter an MFA token or may
be slower logging in. I know we can zoom through the login process in
dev but we don't really need to enter any credentials. Honestly it
_seems_ reasonable to me to increase it further (maybe to five minutes?)
but one minute to two minutes is probably sufficient for now.