Add prometheus counters to track oauth login flow
Add one counter to track number of users that have started the OAuth login
flow. Also add a counter to track the number of users that have finished
the OAuth login process. Together we should be able to get an idea about
how many users are failing to complete the login process.